rasp/IT8951/requirements.txt
RPi.GPIO
pillow
